navmazing collections: step 1 - generic all target for navigation
